[WIP] large refactor for index and add kdtree
