+++ tests/test_suites.c
@@ -14,6 +14,7 @@ int main(void) {
     if (test_round){
         RUN_TEST(test_index);
         RUN_TEST(test_rstar3d);
+        RUN_TEST(test_kdtree2d);
         RUN_TEST(test_geotransfer);
         RUN_TEST(test_readHDF5);
         RUN_TEST(test_interpolate);
