+++ tests/test_suites.h
@@ -9,3 +9,4 @@ void test_interpolate(void);
 void test_comprehensive(void);
 void test_index(void);
 void test_rstar3d(void);
+void test_kdtree2d(void);
